requirements.txt
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
plotly>=5.15.0
flask>=2.3.0
prophet>=1.1.4
scikit-learn>=1.3.0
python-dotenv>=1.0.0
schedule>=1.2.0
